Create qupath.license-conventions
Further shortens build script for qupath-app
